+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 - Add `pauli_propagation` function for efficient approximate simulation of Pauli operators in Heisenberg picture.
 
--  Add `zeros_like` and `ones_like` method for backends.
+- Add `zeros_like`, `ones_like`, `repeat`, `popc`, `top_k`, `lexsort` method for backends.
 
 ### Fixed